requirements.txt
pandas
jupyterlab
spacy
discord.py
python-dotenv
thefuzz